+++ packaging/linux/install-telachat.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env sh
 set -eu
 
-TELACHAT_VERSION="0.63.0"
+TELACHAT_VERSION="0.64.0"
 DEFAULT_RELEASE_URL="https://github.com/H234598/telachat/releases/download/v${TELACHAT_VERSION}/telachat.pyz"
 
 usage() {